+++ src/views/ApplyPage/components/ApplyHeader/index.tsx
@@ -11,7 +11,9 @@ const ApplyHeader = () => {
         <Button buttonStyle="line" padding="10x24">
           임시저장
         </Button>
-        <Button padding="10x24">제출하기</Button>
+        <Button padding="10x24" type="submit">
+          제출하기
+        </Button>
       </div>
     </header>
   );
